chore: amount payable & weight public view fns
